requirements.txt
google-api-python-client==2.127.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.2
google-cloud-logging==3.9.0
python-dotenv==1.0.1
pandas==2.2.0
isodate==0.6.1
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0